Removed get_lowest_day function, Created combinations functions
Realized the best approach to achieve the problem is to use combinations. The combination function takes the array of stocks, iterates over a combination of pairs. Filters out combinations where the first stock price is greather than the second stock price. Another filter would be if the index the first value is greater than the second value it will be removed